moved score recording to after score calculation
