+++ Source/adventureDAO.js
@@ -21,7 +21,6 @@ const { clearBlock, removeModifier } = require("./combatantDAO.js");
 const { spawnEnemy } = require("./enemyDAO.js");
 const { getGuild } = require("./guildDAO.js");
 const { resolveMove } = require("./moveDAO.js");
-const { setPlayer, getPlayer } = require("./playerDAO.js");
 const { renderRoom, updateRoomHeader } = require("./roomDAO.js");
 const { getEquipmentProperty } = require("./equipment/_equipmentDictionary.js");
 
@@ -468,31 +467,7 @@ exports.fetchRecruitMessage = async function (thread, messageId) {
  * @param {string?} descriptionOverride
  */
 exports.completeAdventure = function (adventure, thread, endState, descriptionOverride) {
-	const { guildId, messages: messageManager } = thread;
-	const guildProfile = getGuild(guildId);
-	if (adventure.accumulatedScore > guildProfile.highScore.score) {
-		guildProfile.highScore = {
-			score: adventure.accumulatedScore,
-			playerIds: adventure.delvers.map(delver => delver.id),
-			adventure: adventure.name
-		};
-	}
-	adventure.delvers.forEach(delver => {
-		if (endState !== "giveup") {
-			const player = getPlayer(delver.id, guildId);
-			if (player.scores[guildId]) {
-				player.scores[guildId].total += adventure.accumulatedScore;
-				if (adventure.accumulatedScore > player.scores[guildId].high) {
-					player.scores[guildId].high = adventure.accumulatedScore;
-				}
-			} else {
-				player.scores[guildId] = { total: adventure.accumulatedScore, high: adventure.accumulatedScore };
-			}
-			setPlayer(player);
-		}
-		guildProfile.adventuring.delete(delver.id);
-	})
-
+	const { messages: messageManager } = thread;
 	exports.fetchRecruitMessage(thread, adventure.messageIds.recruit).then(recruitMessage => {
 		const [{ data: recruitEmbed }] = recruitMessage.embeds;
 		recruitMessage.edit({
@@ -508,7 +483,7 @@ exports.completeAdventure = function (adventure, thread, endState, descriptionOv
 	clearComponents(adventure.messageIds.room, messageManager);
 	[adventure.messageIds.utility, adventure.messageIds.deploy, adventure.messageIds.leaderNotice].forEach(id => {
 		if (id) {
-			messageManager.delete(id);
+			messageManager.delete(id).catch(console.error);
 		}
 	})
